requirements.txt
APScheduler==3.11.2
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.61.1
fpdf2==2.8.5
greenlet==3.3.1
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
openpyxl==3.1.5
pandas==2.2.3
pillow==12.1.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
pycparser==3.0
pyOpenSSL==25.3.0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.9
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
signxml==4.2.2
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
WooCommerce==3.0.0
zeep==4.3.2
gunicorn==21.2.0
weasyprint==62.3
